Update and rename classification_wrapper.py to training_modules.py
Leveraging inheritance to avoid duplicating code for different tasks
